requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cssselect==1.3.0
curl_cffi==0.10.0
DataRecorder==3.6.2
ddddocr==1.5.6
DownloadKit==2.0.7
DrissionPage==4.1.0.17
et_xmlfile==2.0.0
filelock==3.17.0
flatbuffers==25.2.10
humanfriendly==10.0
idna==3.10
lxml==5.3.1
mpmath==1.3.0
numpy==2.2.4
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
protobuf==6.30.1
psutil==7.0.0
pycparser==2.22
pycryptodome==3.21.0
PyExecJS2==1.6.1
requests==2.32.3
requests-file==2.1.0
six==1.17.0
sympy==1.13.3
tldextract==5.1.3
urllib3==2.3.0
websocket-client==1.8.0

curl-cffi~=0.10.0
pandas~=2.2.3